backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.6
psycopg2==2.9.6